cover: refact -> add components file
